pyproject.toml
[project]
name = "tp-auth-serverside"
version = "0.1.0"
description = "A server side authentication utility which stores session tokens in memory db."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cryptography>=45.0.5",
    "fastapi>=0.116.1",
    "grpc-interceptor>=0.15.4",
    "grpcio>=1.75.0",
    "httpx>=0.28.1",
    "mem-db-utils>=0.2.0",
    "orjson>=3.11.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "shortuuid>=1.0.13",
]

[build-system]
requires = ["uv_build>=0.8.0,<0.9"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "grpcio-tools>=1.75.0",
]

[[tool.uv.index]]
name = "prismatica-pypi"
url = "https://readUser:hh(iBaf71$icV63@pypi.prismatica.in/simple"

[tool.ruff]
lint.select = [
    "E", # pycodestyle errors
    "W", # pycodestyle warnings
    "F", # pyflakes
    "I", # isort
    "C", # flake8-comprehensions
    "B", # flake8-bugbear
]
lint.ignore = [
    "E501", # line too long, handled by black
    "B008", # do not perform function calls in argument defaults
    "C901", # too complex
    "E402",
    "B904",
    "B905",
    "B009",
    "C417"

]
line-length = 120
target-version = "py312"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true


[tool.mypy]
warn_unused_ignores = true
warn_redundant_casts = true
show_error_codes = true
disallow_untyped_defs = true
ignore_missing_imports = true
follow_imports = "silent"

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
check_untyped_defs = true


[tool.coverage.report]
precision = 2
fail_under = 81.75
show_missing = true
skip_covered = true
exclude_lines = [
    "pragma: no cover",
    "pragma: nocover",
    "if TYPE_CHECKING:",
    "if typing.TYPE_CHECKING:",
    "raise NotImplementedError",
]
